moved login and signup function into user controller, moved file upload functon into user controller
