+++ client/src/api/getUserData.js
@@ -1,0 +1,9 @@
+import axiosAPI from './baseURL';
+
+export const getUserData = (id) => {
+  return axiosAPI.get(`/user/${id}`).then((response) => {
+    return response;
+  }).catch((err) => {
+    console.log(err);
+  })
+};
