+++ client/src/sass/components/Header.module.scss
@@ -48,4 +48,4 @@ header {
       }
     }
   }
-}
+};
