+++ package.json
@@ -33,7 +33,7 @@
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.9.6",
     "morgan": "^1.10.0",
-    "universal-cookie": "^4.0.3"
+    "universal-cookie": "^4.0.3",
     "multer": "^1.4.2"
   },
   "devDependencies": {
