+++ routes/api/login.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 // Controller
-const { login } = require('./../../controllers/authentication_controller');
+const { login } = require('./../../controllers/user_controller');
 
 router.post('/', login);
